Relocate reserved identifier loader into resources
Move src/plugin/src/reserved-identifiers.js into src/plugin/src/resources so the metadata loader lives alongside other plugin data helpers and away from the root surface. Update imports, keep a shim at the old path, and point tests at the new location.
